backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# HTTP client (async)
httpx>=0.26.0

# OpenAI
openai>=1.0.0

# Self-Improving Optimizer
apscheduler>=3.10.0    # Daily improvement cycle scheduling

# For scheduled tasks (optional - see MANUAL_STEPS.md)
# celery>=5.3.0
# redis>=5.0.0
